+++ lib/datadog/core/transport/http/env.rb
